refactor(transform): isolate snapshot value analysis
